singer_project/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
sqlparse==0.4.2
